# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bob - Entrepreneur Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all business opportunities
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [comments]
      summary: List comments for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [comments]
      summary: Add a comment to a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"
...
  /agent/research:
    post:
      operationId: triggerResearch
      tags: [agents]
      summary: Trigger the AI research agent to find business ideas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerResearchBody"
      responses:
        "200":
          description: Research triggered, returns run ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunStatus"
  /agent/orchestrate/{businessId}:
    post:
      operationId: triggerOrchestrate
      tags: [agents]
      summary: Trigger the orchestrator agent to plan/update tasks for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Orchestration triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunStatus"
...
  /agent/runs/{id}:
    get:
      operationId: getAgentRun
      tags: [agents]
      summary: Get status of a specific agent run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        marketSize:
          type: string
        tam:
          type: string
        investmentNeeded:
          type: string
        targetRevenue30d:
          type: string
        effortLevel:
          type: string
          enum: [low, medium, high]
        platform:
          type: string
        status:
          type: string
          enum: [researching, planning, active, completed, paused]
        rank:
          type: integer
        costBenefitScore:
          type: ["number", "null"]
        marketDemandScore:
          type: ["number", "null"]
        agentNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - marketSize
        - tam
        - investmentNeeded
        - targetRevenue30d
        - effortLevel
        - platform
        - status
        - rank
        - costBenefitScore
        - marketDemandScore
        - agentNotes
        - createdAt
        - updatedAt
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        marketSize:
          type: string
        tam:
          type: string
        investmentNeeded:
          type: string
        targetRevenue30d:
          type: string
        effortLevel:
          type: string
          enum: [low, medium, high]
        platform:
          type: string
        rank:
          type: integer
        costBenefitScore:
          type: ["number", "null"]
        marketDemandScore:
          type: ["number", "null"]
        agentNotes:
          type: ["string", "null"]
      required:
        - name
        - description
        - marketSize
        - tam
        - investmentNeeded
        - targetRevenue30d
        - effortLevel
        - platform
        - rank
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        marketSize:
          type: string
        tam:
          type: string
        investmentNeeded:
          type: string
        targetRevenue30d:
          type: string
        effortLevel:
          type: string
          enum: [low, medium, high]
        platform:
          type: string
        status:
          type: string
          enum: [researching, planning, active, completed, paused]
        rank:
          type: integer
        costBenefitScore:
          type: ["number", "null"]
        marketDemandScore:
          type: ["number", "null"]
        agentNotes:
          type: ["string", "null"]
...
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        author:
          type: string
          enum: [user, orchestrator, agent]
        content:
          type: string
        agentType:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - taskId
        - author
        - content
        - agentType
        - createdAt
    CreateTaskCommentBody:
      type: object
      properties:
        author:
          type: string
          enum: [user, orchestrator, agent]
        content:
          type: string
        agentType:
          type: ["string", "null"]
      required:
        - author
        - content
...
    TriggerResearchBody:
      type: object
      properties:
        prompt:
          type: ["string", "null"]
      required: []
    AgentRunStatus:
      type: object
      properties:
        runId:
          type: integer
        status:
          type: string
      required:
        - runId
        - status
    AgentRun:
      type: object
      properties:
        id:
          type: integer
        agentType:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        businessId:
          type: ["integer", "null"]
        log:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agentType
        - status
        - businessId
        - log
        - result
        - createdAt
        - updatedAt
...