# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bob - Entrepreneur Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessEffortLevel } from "./businessEffortLevel";
import type { BusinessStatus } from "./businessStatus";

export interface Business {
  id: number;
  name: string;
  description: string;
  marketSize: string;
  tam: string;
  investmentNeeded: string;
  targetRevenue30d: string;
  effortLevel: BusinessEffortLevel;
  platform: string;
  status: BusinessStatus;
  rank: number;
  /** @nullable */
  costBenefitScore: number | null;
  /** @nullable */
  marketDemandScore: number | null;
  /** @nullable */
  agentNotes: string | null;
  createdAt: string;
  updatedAt: string;
}