# lib/api-zod/src/generated/types/createBusinessBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bob - Entrepreneur Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBusinessBodyEffortLevel } from "./createBusinessBodyEffortLevel";

export interface CreateBusinessBody {
  name: string;
  description: string;
  marketSize: string;
  tam: string;
  investmentNeeded: string;
  targetRevenue30d: string;
  effortLevel: CreateBusinessBodyEffortLevel;
  platform: string;
  rank: number;
  /** @nullable */
  costBenefitScore?: number | null;
  /** @nullable */
  marketDemandScore?: number | null;
  /** @nullable */
  agentNotes?: string | null;
}