# lib/api-zod/src/generated/types/taskComment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bob - Entrepreneur Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskCommentAuthor } from "./taskCommentAuthor";

export interface TaskComment {
  id: number;
  taskId: number;
  author: TaskCommentAuthor;
  content: string;
  /** @nullable */
  agentType: string | null;
  createdAt: string;
}